IGNITE-10561: Fix rebalance iterator issue.

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/tree/io/DataPageIO.java
@@ -303,7 +303,7 @@ public class DataPageIO extends AbstractDataPageIO<CacheDataRow> {
 
         int opCntr = rawNewMvccOperationCounter(addr, 0);
 
-        rawNewMvccOperationCounter(addr, 0, (opCntr & ~MVCC_HINTS_MASK) | (txState << MVCC_HINTS_BIT_OFF));
+        rawNewMvccOperationCounter(addr, 0, (opCntr & ~MVCC_HINTS_MASK) | ((int)txState << MVCC_HINTS_BIT_OFF));
     }
 
     /**
